% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captions.R
\name{cleanCaptionData}
\alias{cleanCaptionData}
\title{Clean up caption data}
\usage{
cleanCaptionData(captionData, removeNonSpokenAnnotations = TRUE)
}
\arguments{
\item{captionData}{Caption data from prior call to \link{getCaptionData}}

\item{removeNonSpokenAnnotations}{Remove non-spoken annotations (default TRUE)}
}
\value{
Returns a \code{\link{data.frame}} of results
}
\description{
Clean up caption data, by removing:
    \itemize{
      \item{inline tags and annotations}
      \item{timing indicators for phrases}
      \item{initial header information}
      \item{styling instructions}
      \item{content start tag}
      \item{extra newlines}
    }

    and, optionally remove non-spoken content annotation is in square brackets or parenthesis

  \href{https://developers.google.com/youtube/v3/docs/captions/list}{https://developers.google.com/youtube/v3/docs/captions/list}
}
\examples{
\dontrun{
cleanData = cleanCaptionData(captionData)
}

}
