% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captions.R
\name{getCaptionData}
\alias{getCaptionData}
\title{Get caption details for a video}
\usage{
getCaptionData(token, contentOwner, captionInfo, format = "vtt")
}
\arguments{
\item{token}{Access token}

\item{contentOwner}{Content owner}

\item{captionInfo}{Caption info from prior call to \link{getVideoCaptionsList}}

\item{format}{format}
}
\value{
Returns a \code{\link{data.frame}} of results
}
\description{
Get caption details for a video

  \href{https://developers.google.com/youtube/v3/docs/captions/list}{https://developers.google.com/youtube/v3/docs/captions/list}
}
\examples{
\dontrun{
message("Retrieving caption info for each video ...")
captions <- videos \%>\% rowwise() \%>\% do(getVideoCaptionsList(token, .$video)) \%>\%
  filter(lang == "en", type \%in\% c("ASR", "standard"))

message("Retrieving captions for each video ...")
captions <- captions \%>\% rowwise() \%>\%
  do(getCaptionData(token = token, contentOwner = "ContentOwner", caption = .))
}

}
