% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChannelMetrics.R
\name{getChannelMetrics}
\alias{getChannelMetrics}
\title{Get metrics for a list of channels}
\usage{
getChannelMetrics(
  token,
  contentOwner,
  from,
  to,
  channels,
  metrics = NULL,
  dimensions = NULL
)
}
\arguments{
\item{token}{Access token}

\item{contentOwner}{Content owner.}

\item{from}{Starting date.}

\item{to}{Ending date.}

\item{channels}{List of channels.}

\item{metrics}{List of metrics. Defaults to \code{c("views")}.}

\item{dimensions}{List of dimensions. Defaults to \code{c("channel")}.}
}
\value{
Returns a \code{\link{data.frame}} of results
}
\description{
Get metrics for a list of channels

  \href{https://developers.google.com/youtube/analytics/content_owner_reports}{https://developers.google.com/youtube/analytics/content_owner_reports}
}
\examples{
\dontrun{
channelMetrics <- getChannelMetrics(token, contentOwner = "ContentOwner",
                                    from = startDate, to = endDate, channels$channelId)
}

}
