% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChannels.R
\name{getChannels}
\alias{getChannels}
\title{Get list of channels controlled by the \code{contentOwner}}
\usage{
getChannels(token, contentOwner)
}
\arguments{
\item{token}{Access token}

\item{contentOwner}{Content owner}
}
\value{
Returns a \code{\link{data.frame}} of results
}
\description{
Get list of channels controlled by the \code{contentOwner}

  \href{https://developers.google.com/youtube/v3/docs/channels/list}{https://developers.google.com/youtube/v3/docs/channels/list}
}
\examples{
\dontrun{
channels <- getChannels(token, contentOwner = "ContentOwner")
}

}
