% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComments.R
\name{getComments}
\alias{getComments}
\title{Get comments below a video}
\usage{
getComments(Vid, api_key)
}
\arguments{
\item{Vid}{Video ID}

\item{api_key}{Your API key}
}
\value{
That'll give you all the comments with their IDs ... you can take those IDs, then, and run them through the comments->list endpoint to get more granular info.
}
\description{

}

