% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOwnerMetrics.R
\name{getOwnerMetrics}
\alias{getOwnerMetrics}
\title{Get metrics for an owner}
\usage{
getOwnerMetrics(
  token,
  contentOwner,
  from,
  to,
  metrics = NULL,
  dimensions = NULL
)
}
\arguments{
\item{token}{Access token}

\item{contentOwner}{Content owner.}

\item{from}{Starting date.}

\item{to}{Ending date.}

\item{metrics}{List of metrics. Defaults to \code{c("views")}.}

\item{dimensions}{List of dimensions. Defaults to \code{c("month")}.}
}
\value{
Returns a \code{\link{data.frame}} of results
}
\description{
Get metrics for an owner

  \href{https://developers.google.com/youtube/analytics/content_owner_reports}{https://developers.google.com/youtube/analytics/content_owner_reports}
}
\examples{
\dontrun{
ownerMetrics <- getOwnerMetrics(token, contentOwner = "ContentOwner",
                                from = startDate, to = endDate)
}

}
