% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToken.R
\name{getServiceAccountToken}
\alias{getServiceAccountToken}
\title{Get service account token}
\usage{
getServiceAccountToken(service_token_creds_file)
}
\arguments{
\item{service_token_creds_file}{Credentials file}
}
\value{
token A service account token
}
\description{
Authenticate and generate service account token
}
\examples{
\dontrun{
# Credentials with "content owner" role, with access to all the needed YT channels
token <- getServiceAccountToken("ServiceAccount-e8a41d405bf8.json")
}

}
