% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captions.R
\name{getVideoCaptionsList}
\alias{getVideoCaptionsList}
\title{Get list of captions for a video}
\usage{
getVideoCaptionsList(token, videoId)
}
\arguments{
\item{token}{Access token}

\item{videoId}{Video ID}
}
\value{
Returns a \code{\link{data.frame}} of results
}
\description{
Get list of captions for a video

  \href{https://developers.google.com/youtube/v3/docs/captions/list}{https://developers.google.com/youtube/v3/docs/captions/list}
}
\examples{
\dontrun{
captionList <- getVideoCaptionsList(token, video)
}

}
