% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVideoInfo.R
\name{getVideoInfo}
\alias{getVideoInfo}
\title{Get information on a list of videos}
\usage{
getVideoInfo(token, videos)
}
\arguments{
\item{token}{Access token}

\item{videos}{List of videos}
}
\value{
Returns a \code{\link{data.frame}} of results
}
\description{
Get information on a list of videos

  \href{https://developers.google.com/youtube/v3/docs/videos/list}{https://developers.google.com/youtube/v3/docs/videos/list}
}
\examples{
\dontrun{
videoInfo <- getVideoInfo(token, videos$video)
}

}
