% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVideoList.R
\name{getVideoList}
\alias{getVideoList}
\title{Get list of videos from a channel}
\usage{
getVideoList(token, playlistId, limit = 50)
}
\arguments{
\item{token}{Access token}

\item{playlistId}{Playlist ID}

\item{limit}{Number of videos to return. Set to \code{Inf} for all}
}
\value{
Returns a \code{\link{data.frame}} of results
}
\description{
Get list of videos associated with a \code{playlistId}

  \href{https://developers.google.com/youtube/v3/docs/playlistItems/list}{https://developers.google.com/youtube/v3/docs/playlistItems/list}
}
\examples{
\dontrun{
videos <- getVideoList(token, playlistId = ch$uploadsPlaylist, limit = Inf)
}

}
