% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVideoPlaythrough.R
\name{getVideoPlaythrough}
\alias{getVideoPlaythrough}
\title{Get play-through metrics for a list of videos}
\usage{
getVideoPlaythrough(token, contentOwner, from, to, videos, metrics = NULL)
}
\arguments{
\item{token}{Access token}

\item{contentOwner}{Content owner}

\item{from}{Starting date}

\item{to}{Ending date}

\item{videos}{List of videos}

\item{metrics}{List of audience retention metrics. Defaults to \code{c("audienceWatchRatio")}}
}
\value{
Returns a \code{\link{data.frame}} of results
}
\description{
Get audience retention, a.k.a. play-through metrics for a list of videos

  \href{https://developers.google.com/youtube/analytics/metrics#Audience_Retention_Metrics}{https://developers.google.com/youtube/analytics/metrics#Audience_Retention_Metrics}
}
\examples{
\dontrun{
videoMetrics <- getVideoPlaythrough(token, contentOwner = "ContentOwner",
                               from = "2019-02-03", to = "2019-02-09", videos$video)
}

}
