% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RYoutube-package.R
\docType{package}
\name{RYoutbe-package}
\alias{RYouTube}
\alias{RYoutbe-package}
\title{Process Youtube data using  R}
\description{
This package provides a series of functions that allow R users
to access Youtube's API to get information about channels and videos
}
\author{
Saber Bouabid \email{saber@bi-statistics.com}
}
\seealso{
, \code{\link{geTrends}},
\code{\link{getComments}}, \code{\link{Yousearch}},
\code{\link{uploadYT}}, \code{\link{getStats}},
\code{\link{getVideos}}, \code{\link{getChannel}}
}

